mf_sil/contrib/Reporting_UI/pl_parking/requirements.txt
adas-tsf>=2.4.1
asammdf
drawsvg
fiona
geopandas
numpy
opencv-python
plotly
plotly-express
psycopg2-binary
pyproj
rrecreader
scikit-learn
torch
